+++ src/Widget/Settings/SettingsWindow.cpp
@@ -23,7 +23,7 @@ SettingsWindow::SettingsWindow( QWidget *parent ) :
 
 	initMenu();
     initWidgets();
-    showSettings( ui->treeView->model()->index( 0, 0 ) );
+    showSettings( ui->treeView->model()->index( 0, 1 ) );
 
     connect( ui->leFilter, SIGNAL( textChanged( const QString& ) ), this, SLOT( filterMenu( const QString& ) ) );
     connect( ui->treeView, SIGNAL( clicked( const QModelIndex& ) ), this, SLOT( showSettings( const QModelIndex& ) ) );
@@ -39,8 +39,8 @@ SettingsWindow::~SettingsWindow()
 
 void SettingsWindow::initWidgets()
 {
-	widgets["General"]	= new SettingsWidgetGeneral();
-	widgets["Speaker"]	= new SettingsWidgetSpeaker();
+	widgets["general"]	= new SettingsWidgetGeneral();
+	widgets["speaker"]	= new SettingsWidgetSpeaker();
 
 	foreach ( AbstractSettingsWidget* wdg, widgets ) {
 		ui->mainWidget->addWidget( wdg );
@@ -49,26 +49,22 @@ void SettingsWindow::initWidgets()
 
 void SettingsWindow::initMenu()
 {
-	/* TREE EXAMPLE
-	QFile file( ":/Resources/settings_menu/menu.txt" );
-	TreeModel* model	= new TreeModel( file.readAll(), this, Txt );
-	*/
 	QFile file( ":/Resources/settings_menu/menu.xml" );
 	file.open( QIODevice::ReadOnly );
-	TreeModel* sourceModel	= new TreeModel( file.readAll(), this, Xml );
+	TreeModel* sourceModel	= new TreeModel( file.readAll(), this );
 	menuModel				= new QSortFilterProxyModel( this );
 
 	menuModel->setSourceModel( sourceModel );
 	menuModel->setFilterKeyColumn( 1 );
 
 	ui->treeView->setModel( menuModel );
-	ui->treeView->hideColumn( 0 );	// Id Column for this model
+	ui->treeView->hideColumn( 1 );	// Id Column for this model
 }
 
 void SettingsWindow::showSettings( const QModelIndex &index )
 {
-	QString id		= ui->treeView->model()->data( index.siblingAtColumn( 0 ) ).toString();
-	QString title	= ui->treeView->model()->data( index.siblingAtColumn( 1 ) ).toString();
+	QString title	= ui->treeView->model()->data( index.siblingAtColumn( 0 ) ).toString();
+	QString id		= ui->treeView->model()->data( index.siblingAtColumn( 1 ) ).toString();
 	//qDebug() << "TreeView ID: " << id;
 	showWidget( id, title );
 }
@@ -88,7 +84,7 @@ void SettingsWindow::applySettings()
 {
 	AbstractSettingsWidget*	w	= qobject_cast<AbstractSettingsWidget*>( ui->mainWidget->currentWidget() );
 	w->apply();
-	if ( w == widgets["Speaker"] )
+	if ( w == widgets["speaker"] )
 		emit speakerSettingsUpdated();
 }